Initial commit for feature request #2. - Offers solution for steps 1-3
